requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
httpx
pydantic
pydantic-settings
